# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: PS Parts API
  version: 0.0.0
tags: []
paths:
  /api/auth/me:
    get:
      operationId: currentUser
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
  /api/auth/signin:
    post:
      operationId: signin
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
  /api/auth/signout:
    post:
      operationId: signout
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                required:
                  - success
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/signup:
    post:
      operationId: signup
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
  /api/categories:
    get:
      operationId: listCategories
      description: Список категорий, отсортированный по id.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/products:
    get:
      operationId: listProducts
      description: Каталог товаров с названиями категорий, отсортированный по id.
      parameters:
        - name: category
          in: query
          required: false
          description: Необязательный фильтр по slug категории.
          schema:
            type: string
            pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
            minLength: 1
            maxLength: 100
          explode: false
        - name: priceMin
          in: query
          required: false
          description: Минимальная цена в целых рублях, включительно.
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 21474836
          explode: false
        - name: priceMax
          in: query
          required: false
          description: Максимальная цена в целых рублях, включительно.
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 21474836
          explode: false
        - name: available
          in: query
          required: false
          schema:
            type: boolean
          explode: false
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
          explode: false
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100000
            default: 1
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 48
            default: 12
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/products/{id}:
    get:
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /health-check:
    get:
      operationId: healthCheck
      description: Проверка доступности сервера.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
components:
  schemas:
    ApiError:
      type: object
      required:
        - statusCode
        - error
        - message
      properties:
        statusCode:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        error:
          type: string
        message:
          type: string
      description: Единый JSON-формат ошибок API. statusCode совпадает с HTTP-статусом.
      examples:
        - statusCode: 400
          error: Bad Request
          message: Некорректный запрос
    Category:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: integer
          format: int32
        slug:
          type: string
        name:
          type: string
    Credentials:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          minLength: 3
          maxLength: 254
          pattern: ^[^\s@]+@[^\s@]+\.[^\s@]+$
        password:
          type: string
          minLength: 8
          maxLength: 128
    Health:
      type: object
      required:
        - health
      properties:
        health:
          type: string
          enum:
            - check
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
          format: int32
          minimum: 0
        currency:
          type: string
          enum:
            - RUB
      description: 'Денежная сумма: amount — целое число копеек, currency — RUB.'
      examples:
        - amount: 1999000
          currency: RUB
    Product:
      type: object
      required:
        - id
        - sku
        - name
        - description
        - price
        - category_slug
        - category_name
        - image_url
        - available
      properties:
        id:
          type: integer
          format: int32
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        category_slug:
          type: string
        category_name:
          type: string
        image_url:
          anyOf:
            - type: string
            - type: 'null'
        available:
          type: boolean
    ProductPage:
      type: object
      required:
        - items
        - total
        - page
        - pageSize
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          format: int32
          minimum: 0
        page:
          type: integer
          format: int32
          minimum: 1
        pageSize:
          type: integer
          format: int32
          minimum: 1
        totalPages:
          type: integer
          format: int32
          minimum: 0
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: psparts_session